+++ src/reuse/header.py
@@ -14,10 +14,11 @@ import datetime
 import logging
 import re
 import sys
+from argparse import ArgumentParser
 from gettext import gettext as _
 from os import PathLike
 from pathlib import Path
-from typing import NamedTuple, Optional, Sequence
+from typing import List, NamedTuple, Optional, Sequence
 
 from binaryornot.check import is_binary
 from boolean.boolean import ParseError
@@ -281,32 +282,27 @@ def find_and_replace_header(
     return new_text
 
 
-def _get_comment_style(path: Path) -> CommentStyle:
-    """Return value of CommentStyle detected for *path*.
-
-    :raises KeyError: if no comment style is detected.
-    """
+def _get_comment_style(path: Path) -> Optional[CommentStyle]:
+    """Return value of CommentStyle detected for *path* or None."""
     style = FILENAME_COMMENT_STYLE_MAP.get(path.name)
     if style is None:
-        style = EXTENSION_COMMENT_STYLE_MAP[path.suffix]
+        style = EXTENSION_COMMENT_STYLE_MAP.get(path.suffix)
     return style
 
 
-def _verify_paths_supported(
-    paths, parser, force_single=False, force_multi=False
+def _verify_paths_line_handling(
+    paths: List[Path],
+    parser: ArgumentParser,
+    force_single: bool,
+    force_multi: bool,
 ):
+    """This function aborts the parser when *force_single* or *force_multi* is
+    used, but the file type does not support that type of comment style.
+    """
     for path in paths:
-        try:
-            style = _get_comment_style(path)
-        except KeyError:
-            # TODO: This check is duplicated.
-            if not is_binary(str(path)):
-                parser.error(
-                    _(
-                        "'{path}' does not have a recognised file extension,"
-                        " please use --style or --explicit-license"
-                    ).format(path=path)
-                )
+        style = _get_comment_style(path)
+        if style is None:
+            continue
         if force_single and not style.can_handle_single():
             parser.error(
                 _(
@@ -319,6 +315,20 @@ def _verify_paths_supported(
                 _(
                     "'{path}' does not support multi-line comments, please"
                     " do not use --multi-line"
+                ).format(path=path)
+            )
+
+
+def _verify_paths_comment_style(paths: List[Path], parser: ArgumentParser):
+    for path in paths:
+        style = _get_comment_style(path)
+        # TODO: This check is duplicated.
+        if style is None and not is_binary(str(path)):
+            parser.error(
+                _(
+                    "'{path}' does not have a recognised file extension,"
+                    " please use --style, --explicit-license or"
+                    " --skip-unrecognised"
                 ).format(path=path)
             )
 
@@ -363,6 +373,10 @@ def _add_header_to_file(
         style = NAME_STYLE_MAP[style]
     else:
         style = _get_comment_style(path)
+        if style is None:
+            out.write(_("Skipped unrecognised file {path}").format(path=path))
+            out.write("\n")
+            return result
 
     with path.open("r") as fp:
         text = fp.read()
@@ -460,6 +474,11 @@ def add_arguments(parser) -> None:
         action="store_true",
         help=_("place header in path.license instead of path"),
     )
+    parser.add_argument(
+        "--skip-unrecognised",
+        action="store_true",
+        help=_("skip files with unrecognised comment styles"),
+    )
     parser.add_argument("path", action="store", nargs="+", type=PathType("w"))
 
 
@@ -479,16 +498,26 @@ def run(args, project: Project, out=sys.stdout) -> int:
             _("option --single-line and --multi-line are mutually exclusive")
         )
 
+    if args.style is not None and args.skip_unrecognised:
+        _LOGGER.warning(
+            _(
+                "--skip-unrecognised has no effect when used together with"
+                " --style"
+            )
+        )
+
     paths = [_determine_license_path(path) for path in args.path]
 
-    # First loop to verify before proceeding
+    # Verify line handling and comment styles before proceeding
     if args.style is None and not args.explicit_license:
-        _verify_paths_supported(
+        _verify_paths_line_handling(
             paths,
             args.parser,
             force_single=args.single_line,
             force_multi=args.multi_line,
         )
+        if not args.skip_unrecognised:
+            _verify_paths_comment_style(paths, args.parser)
 
     template = None
     commented = False